Remove unneeded VLAN stuff and clean up CMR in LinodeDetail
