+++ packages/manager/src/features/linodes/LinodesDetail/LinodeSummary/NetworkGraph.tsx
@@ -23,7 +23,7 @@ import {
 } from 'src/utilities/statMetrics';
 import { readableBytes } from 'src/utilities/unitConversions';
 import StatsPanel from './StatsPanel';
-import TotalTraffic, { TotalTrafficProps } from './TotalTraffic';
+import { TotalTrafficProps } from './TotalTraffic';
 import { ChartProps } from './types';
 
 const formatTotalTraffic = (value: number) =>
@@ -32,10 +32,15 @@ const formatTotalTraffic = (value: number) =>
 const useStyles = makeStyles((theme: Theme) => ({
   chart: {
     position: 'relative',
-    paddingLeft: theme.spacing(1)
+    paddingTop: theme.spacing(2),
+    paddingLeft: theme.spacing(3)
   },
-  totalTraffic: {
-    margin: '12px'
+  graphGrids: {
+    flexWrap: 'nowrap',
+    paddingLeft: '8px',
+    [theme.breakpoints.down('sm')]: {
+      flexWrap: 'wrap'
+    }
   }
 }));
 
@@ -72,6 +77,8 @@ const _getMetrics = (data: NetworkStats) => {
 
 export const NetworkGraph: React.FC<CombinedProps> = props => {
   const { rangeSelection, stats, theme, ...rest } = props;
+
+  const classes = useStyles();
 
   const v4Data: NetworkStats = {
     publicIn: pathOr([], ['data', 'netv4', 'in'], stats),
@@ -138,34 +145,38 @@ export const NetworkGraph: React.FC<CombinedProps> = props => {
   };
 
   return (
-    <>
-      <StatsPanel
-        title={`IPv4 Traffic (${v4Unit}/s)`}
-        renderBody={() => (
-          <Graph
-            data={v4Data}
-            unit={v4Unit}
-            totalTraffic={v4totalTraffic}
-            metrics={v4Metrics}
-            {...commonGraphProps}
-          />
-        )}
-        {...rest}
-      />
-      <StatsPanel
-        title={`IPv6 Traffic (${v6Unit}/s)`}
-        renderBody={() => (
-          <Graph
-            data={v6Data}
-            unit={v6Unit}
-            totalTraffic={v6totalTraffic}
-            metrics={v6Metrics}
-            {...commonGraphProps}
-          />
-        )}
-        {...rest}
-      />
-    </>
+    <Grid container direction="row" className={classes.graphGrids}>
+      <Grid item xs={12}>
+        <StatsPanel
+          title={`Network — IPv4 (${v4Unit}/s)`}
+          renderBody={() => (
+            <Graph
+              data={v4Data}
+              unit={v4Unit}
+              totalTraffic={v4totalTraffic}
+              metrics={v4Metrics}
+              {...commonGraphProps}
+            />
+          )}
+          {...rest}
+        />
+      </Grid>
+      <Grid item xs={12}>
+        <StatsPanel
+          title={`Network — IPv6 (${v6Unit}/s)`}
+          renderBody={() => (
+            <Graph
+              data={v6Data}
+              unit={v6Unit}
+              totalTraffic={v6totalTraffic}
+              metrics={v6Metrics}
+              {...commonGraphProps}
+            />
+          )}
+          {...rest}
+        />
+      </Grid>
+    </Grid>
   );
 };
 
@@ -189,8 +200,7 @@ const Graph: React.FC<GraphProps> = props => {
     rangeSelection,
     theme,
     timezone,
-    unit,
-    totalTraffic
+    unit
   } = props;
 
   const format = formatBitsPerSecond;
@@ -216,71 +226,60 @@ const Graph: React.FC<GraphProps> = props => {
   const convertedPrivateOut = data.privateOut;
 
   return (
-    <React.Fragment>
-      <div className={classes.chart}>
-        <LineGraph
-          timezone={timezone}
-          chartHeight={chartHeight}
-          unit={`/s`}
-          formatData={convertNetworkData}
-          formatTooltip={_formatTooltip}
-          showToday={rangeSelection === '24'}
-          data={[
-            {
-              borderColor: 'transparent',
-              backgroundColor: theme.graphs.network.inbound,
-              data: convertedPublicIn,
-              label: 'Public Inbound'
-            },
-            {
-              borderColor: 'transparent',
-              backgroundColor: theme.graphs.network.outbound,
-              data: convertedPublicOut,
-              label: 'Public Outbound'
-            },
-            {
-              borderColor: 'transparent',
-              backgroundColor: theme.graphs.purple,
-              data: convertedPrivateIn,
-              label: 'Private Inbound'
-            },
-            {
-              borderColor: 'transparent',
-              backgroundColor: theme.graphs.yellow,
-              data: convertedPrivateOut,
-              label: 'Private Outbound'
-            }
-          ]}
-          legendRows={[
-            {
-              data: metrics.publicIn,
-              format
-            },
-            {
-              data: metrics.publicOut,
-              format
-            },
-            {
-              data: metrics.privateIn,
-              format
-            },
-            {
-              data: metrics.privateOut,
-              format
-            }
-          ]}
-        />
-      </div>
-      {rangeSelection === '24' && (
-        <Grid item xs={12} lg={6} className={classes.totalTraffic}>
-          <TotalTraffic
-            inTraffic={totalTraffic.inTraffic}
-            outTraffic={totalTraffic.outTraffic}
-            combinedTraffic={totalTraffic.combinedTraffic}
-          />
-        </Grid>
-      )}
-    </React.Fragment>
+    <div className={classes.chart}>
+      <LineGraph
+        timezone={timezone}
+        chartHeight={chartHeight}
+        unit={`/s`}
+        formatData={convertNetworkData}
+        formatTooltip={_formatTooltip}
+        showToday={rangeSelection === '24'}
+        data={[
+          {
+            borderColor: 'transparent',
+            backgroundColor: theme.graphs.network.inbound,
+            data: convertedPublicIn,
+            label: 'Public Inbound'
+          },
+          {
+            borderColor: 'transparent',
+            backgroundColor: theme.graphs.network.outbound,
+            data: convertedPublicOut,
+            label: 'Public Outbound'
+          },
+          {
+            borderColor: 'transparent',
+            backgroundColor: theme.graphs.purple,
+            data: convertedPrivateIn,
+            label: 'Private Inbound'
+          },
+          {
+            borderColor: 'transparent',
+            backgroundColor: theme.graphs.yellow,
+            data: convertedPrivateOut,
+            label: 'Private Outbound'
+          }
+        ]}
+        legendRows={[
+          {
+            data: metrics.publicIn,
+            format
+          },
+          {
+            data: metrics.publicOut,
+            format
+          },
+          {
+            data: metrics.privateIn,
+            format
+          },
+          {
+            data: metrics.privateOut,
+            format
+          }
+        ]}
+      />
+    </div>
   );
 };
